Added button to test alarm in SearchScreenFragment.kt Moved AppIntroActivity.kt to root of ui package
