+++ app/src/main/res/layout/activity_app_intro.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.appintro.AppIntroActivity">
+    tools:context=".ui.AppIntroActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
