+++ app/src/main/res/layout/search_screen_fragment.xml
@@ -5,10 +5,12 @@
     tools:context=".searchscreen.SearchScreenFragment">
 
     <data>
-        <import type="android.view.View"/>
+
+        <import type="android.view.View" />
+
         <variable
             name="searchScreenViewModel"
-            type="com.nachc.dba.searchscreen.SearchScreenViewModel"/>
+            type="com.nachc.dba.searchscreen.SearchScreenViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
@@ -19,15 +21,15 @@
 
         <LinearLayout
             android:id="@+id/searchLinearLayout"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:background="@drawable/custom_search_box"
             android:orientation="horizontal"
-            android:padding="4dp">
+            android:padding="4dp"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent">
 
             <EditText
                 android:id="@+id/inputLineEditText"
@@ -46,7 +48,7 @@
                 android:paddingEnd="12dp"
                 android:paddingBottom="8dp"
                 android:textAlignment="textStart"
-                android:textColor="#000000"/>
+                android:textColor="#000000" />
 
             <Button
                 android:id="@+id/searchBtn"
@@ -56,28 +58,24 @@
                 android:layout_marginStart="4dp"
                 android:layout_marginEnd="2dp"
                 android:background="@drawable/btn_search_custom"
-                android:text="@string/searchBtn"
-                />
+                android:text="@string/searchBtn" />
         </LinearLayout>
 
         <ProgressBar
             android:id="@+id/progressBar"
             style="?android:attr/progressBarStyle"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginBottom="265dp"
+            android:visibility="gone"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintHorizontal_bias="0.5"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/searchLinearLayout"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginBottom="265dp"
-            android:visibility="gone" />
+            app:layout_constraintTop_toBottomOf="@id/searchLinearLayout" />
 
         <TextView
             android:id="@+id/permissionText"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/progressBar"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginStart="177dp"
@@ -86,21 +84,24 @@
             android:text="@string/permission_text"
             android:textColor="#CF0F0F"
             android:textSize="14sp"
-            android:visibility="gone" />
+            android:visibility="gone"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@+id/progressBar" />
 
         <Button
             android:id="@+id/permissionBtn"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintHorizontal_bias="0.5"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/permissionText"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginStart="166dp"
             android:layout_marginTop="27dp"
             android:layout_marginEnd="157dp"
+            android:text="@string/settings_button"
             android:visibility="gone"
-            android:text="@string/settings_button"/>
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintHorizontal_bias="0.5"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@+id/permissionText" />
 
         <Button
             android:id="@+id/allowLocationBtn"
@@ -131,16 +132,41 @@
 
         <TextView
             android:id="@+id/loadError"
+            style="@style/Text"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="@string/routeError"
             android:textColor="@color/design_default_color_error"
-            style="@style/Text"
             android:visibility="gone"
             app:layout_constraintBottom_toTopOf="@+id/dataSavedTextView"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/searchLinearLayout" />
+
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:gravity="center_vertical"
+            app:layout_constraintBottom_toBottomOf="parent">
+
+            <Button
+                android:id="@+id/testAlarmBtn"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_margin="15dp"
+                android:background="@color/colorPrimary"
+                android:text="@string/testAlarmBtn"
+                android:textSize="12sp" />
+
+            <TextView
+                android:id="@+id/testAlarmHelp"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="@string/testAlarmHelp"
+                android:textStyle="bold"
+                android:textColor="@color/colorAccent" />
+
+        </LinearLayout>
 
     </androidx.constraintlayout.widget.ConstraintLayout>
